✅ 5Q6T1E task: split sync contract tests
