+++ packages/core/src/tasks/task-artifact-schema.test.ts
@@ -316,105 +316,6 @@ describe("task-artifact-schema", () => {
     expect(() => validateAcr(validAcr())).not.toThrow();
   });
 
-  it("validates the provider-neutral task sync envelope in task README frontmatter", () => {
-    const frontmatter = withTaskReadmeFrontmatterDefaults({
-      id: "202605291916-5Q6T1E",
-      title: "Synced task",
-      status: "TODO",
-      priority: "high",
-      owner: "CODER",
-      depends_on: [],
-      tags: ["cloud"],
-      verify: ["bun run test"],
-      comments: [],
-      doc_version: 3,
-      doc_updated_at: "2026-05-29T19:16:00.000Z",
-      doc_updated_by: "CODER",
-      description: "Task with external projection metadata.",
-      sync: {
-        version: 1,
-        external_refs: [
-          {
-            provider: "github-projects",
-            connector_kind: "cloud",
-            connection_id: "conn-1",
-            remote_id: "item-1",
-            remote_url: "https://github.com/orgs/example/projects/1/views/1?pane=issue&itemId=1",
-            remote_revision: "rev-1",
-            title: "Remote title",
-            state: "Todo",
-            synced_at: "2026-05-29T19:16:00.000Z",
-          },
-        ],
-        field_policies: {
-          status: {
-            authority: "provider",
-            remote_field: "Status",
-            conflict_policy: "record",
-            updated_at: "2026-05-29T19:16:00.000Z",
-          },
-          "sections.Plan": {
-            authority: "agentplane",
-            conflict_policy: "agentplane_wins",
-          },
-        },
-        freshness: {
-          projected_at: "2026-05-29T19:16:00.000Z",
-          projection_sha256: "sha256:abcdef",
-          source_revision: 3,
-          provider_revision: "rev-1",
-          stale: false,
-        },
-        conflicts: [
-          {
-            id: "conflict-1",
-            kind: "field",
-            severity: "warning",
-            status: "open",
-            summary: "Remote status differs from local task status.",
-            provider: "github-projects",
-            remote_id: "item-1",
-            field: "status",
-            detected_at: "2026-05-29T19:16:00.000Z",
-            safe_command: "agentplane backend sync cloud --direction pull --conflict=diff --yes",
-            when_to_stop: "stop if provider status should not override local status",
-          },
-        ],
-      },
-    });
-
-    expect(() => validateTaskReadmeFrontmatter(frontmatter)).not.toThrow();
-  });
-
-  it("rejects malformed task sync envelope shapes", () => {
-    const frontmatter = withTaskReadmeFrontmatterDefaults({
-      id: "202605291916-5Q6T1E",
-      title: "Synced task",
-      status: "TODO",
-      priority: "high",
-      owner: "CODER",
-      depends_on: [],
-      tags: [],
-      verify: [],
-      comments: [],
-      doc_version: 3,
-      doc_updated_at: "2026-05-29T19:16:00.000Z",
-      doc_updated_by: "CODER",
-      description: "Task with invalid external projection metadata.",
-      sync: {
-        version: 2,
-        external_refs: [{ provider: "jira" }],
-        field_policies: {
-          status: { authority: "jira" },
-        },
-        conflicts: [],
-      },
-    });
-
-    const errors = listTaskReadmeFrontmatterSchemaErrors(frontmatter);
-    expect(errors.join("\n")).toContain("sync");
-  });
-
   it("validates a dry-run AgentPlane runner handoff contract", () => {
     expect(() =>
       validateAgentPlaneRunnerHandoff(validRunnerHandoff(), {
@@ -767,145 +668,6 @@ describe("task-artifact-schema", () => {
         },
       }),
     ).not.toThrow();
-  });
-
-  it("accepts a provider-neutral task sync envelope in task README frontmatter and exports", () => {
-    const task = withTaskReadmeFrontmatterDefaults({
-      id: "202605291916-5Q6T1E",
-      title: "Provider-neutral sync envelope fixture",
-      status: "TODO",
-      priority: "high",
-      owner: "CODER",
-      depends_on: [],
-      tags: ["code"],
-      verify: [],
-      plan_approval: { state: "approved", updated_at: null, updated_by: null, note: null },
-      verification: { state: "pending", updated_at: null, updated_by: null, note: null },
-      comments: [],
-      doc_version: 3,
-      doc_updated_at: "2026-05-29T19:17:18.812Z",
-      doc_updated_by: "CODER",
-      description: "Fixture",
-      id_source: "generated",
-      sync: {
-        version: 1,
-        external_refs: [
-          {
-            provider: "gitlab",
-            connector_kind: "gitlab_project",
-            connection_id: "conn_123",
-            remote_id: "42",
-            remote_url: "https://gitlab.example/group/project/-/issues/42",
-            remote_revision: "rev-1",
-            synced_at: "2026-05-29T19:18:00.000Z",
-          },
-        ],
-        field_policies: {
-          status: {
-            authority: "bidirectional",
-            remote_field: "state",
-            conflict_policy: "record",
-          },
-          verification: {
-            authority: "agentplane",
-            conflict_policy: "agentplane_wins",
-          },
-        },
-        freshness: {
-          projected_at: "2026-05-29T19:18:00.000Z",
-          projection_sha256: "abc123",
-          source_revision: 3,
-          provider_revision: "rev-1",
-          stale: false,
-        },
-        conflicts: [
-          {
-            id: "conflict_1",
-            kind: "field",
-            severity: "warning",
-            status: "open",
-            summary: "Provider status changed after local projection.",
-            provider: "gitlab",
-            remote_id: "42",
-            field: "status",
-            detected_at: "2026-05-29T19:19:00.000Z",
-            safe_command: "agentplane backend sync cloud --direction pull",
-            when_to_stop: "Stop before overwriting local verification evidence.",
-          },
-        ],
-      },
-    });
-
-    expect(() => validateTaskReadmeFrontmatter(task)).not.toThrow();
-    expect(() =>
-      validateTasksExportSnapshot({
-        tasks: [
-          {
-            ...task,
-            commit: null,
-            dirty: false,
-          },
-        ],
-        meta: {
-          schema_version: 1,
-          managed_by: "agentplane",
-          checksum_algo: "sha256",
-          checksum: "abc",
-        },
-      }),
-    ).not.toThrow();
-  });
-
-  it("rejects malformed task sync envelope entries", () => {
-    const errors = listTaskReadmeFrontmatterSchemaErrors(
-      withTaskReadmeFrontmatterDefaults({
-        id: "202605291916-5Q6T1E",
-        title: "Malformed sync envelope fixture",
-        status: "TODO",
-        priority: "high",
-        owner: "CODER",
-        depends_on: [],
-        tags: ["code"],
-        verify: [],
-        plan_approval: { state: "approved", updated_at: null, updated_by: null, note: null },
-        verification: { state: "pending", updated_at: null, updated_by: null, note: null },
-        comments: [],
-        doc_version: 3,
-        doc_updated_at: "2026-05-29T19:17:18.812Z",
-        doc_updated_by: "CODER",
-        description: "Fixture",
-        sync: {
-          version: 1,
-          external_refs: [{ provider: "gitlab" }],
-          field_policies: { status: { authority: "remote" } },
-          conflicts: [
-            {
-              id: "conflict_1",
-              kind: "field",
-              severity: "warning",
-              status: "open",
-              summary: "Missing detected timestamp.",
-            },
-          ],
-        },
-      }),
-    );
-
-    expect(errors).toHaveLength(1);
-    expect(errors[0]).toContain("sync.external_refs[0].remote_id");
-    expect(errors[0]).toContain("sync.field_policies.status.authority");
-    expect(errors[0]).toContain("sync.conflicts[0].detected_at");
-  });
-
-  it("renders task sync envelope fields in the published schema", () => {
-    const schema = JSON.parse(renderTaskReadmeFrontmatterSchemaJson()) as {
-      properties?: Record<string, unknown>;
-    };
-
-    expect(schema.properties).toHaveProperty("sync");
-    expect(JSON.stringify(schema.properties?.sync)).toContain("external_refs");
-    expect(JSON.stringify(schema.properties?.sync)).toContain("field_policies");
-    expect(JSON.stringify(schema.properties?.sync)).toContain("conflicts");
   });
 
   it("accepts context task kinds and context blueprint requests", () => {
